Varied Speed and showCharGui
Character.cs: Made the wd() method's Thread.Sleep depend on the speed setting added and the length of the text. I have reworked the wait feature so that on smaller lines the wait time is less and vice versa for larger lines. Also added a showCharGui method based off of dereks old one. It works and shows all the statistics for when we need to show the player their character. The style is based off of a 'papers please' type identification document.
CharacterCreation.cs: Changed some wording & fixed a console.clear bug.
SettingsMenu.cs: Reworked the speedsetting option to fit with the new system. & replaced old thread.sleeps with working ones.
BoardShip: Changed a readline to pressAnyKeyToContinue()