Tomohawk Giving
ItemID.cs: Added Tomahawk to switch statements
Item.cs: Added showItemInfoGUI for displaying the information of an item (such as before taking or purchasing it)
TowardsMedallion: Made it so that the Tomohawk can be found when discovering it in the barrel.
CharacterCreation.cs: Had to fix some weird conflict stuff that was ambiguous anyway.

Updates some price and damage values in all of the different weapons